requirements.txt
MySQL-python==1.2.4
SQLAlchemy==0.8.1
distribute==0.6.34
nose==1.3.0
python-dateutil==2.1
six==1.3.0
wsgiref==0.1.2
